requirements.txt
cookiecutter==2.6.0
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-htmx==1.21.0
django-tailwind==3.8.0
django-unfold==0.45.0
python-dotenv==1.0.0
stripe==11.4.1
